+++ src/lib.rs
@@ -25,6 +25,7 @@ mod config;
 mod divan;
 mod entry;
 mod stats;
+mod thread_pool;
 mod time;
 mod tree_painter;
 mod util;